refactor: 독립적으로 단위 테스트를 할 수 있도록 인터페이스에 의존하도록 변경
직접 참조하고 있는 BaseballNumberBundleValidator를 인터페이스에 의존하도록 만들어서,
단위 테스트하기 쉽게 변경합니다.
